enemy ship difficulty update

+++ include/game/laser.hpp
@@ -23,7 +23,7 @@ class laser final : public wze::entity {
     float _z_speed;
     float _half_length;
     float _length_skip;
-    float _damage;
+    uint16_t _damage;
     bool _destroyed;
     std::vector<std::shared_ptr<wze::sprite>> _sprites;
 
@@ -31,7 +31,7 @@ class laser final : public wze::entity {
     float z() const;
     laser(float x, float y, float z, float x_speed, float y_speed,
           float z_speed, float length, float diameter, uint8_t color_r,
-          uint8_t color_g, uint8_t color_b, float damage);
+          uint8_t color_g, uint8_t color_b, uint16_t damage);
     bool update(player_ship& player_ship, std::vector<enemy_ship>& enemy_ships,
                 std::vector<asteroid>& asteroids,
                 std::vector<asteroid_loot>& asteroid_loots);
